Refactor labelingPage actions, type, & reducer
