+++ VisionOnEdge/ui/src/store/rootReducer.ts
@@ -4,7 +4,7 @@ import { persistReducer } from 'redux-persist';
 
 import camerasReducer from './camera/cameraReducer';
 import partReducer from './part/partReducer';
-import labelingPageStateReducer from '../reducer/labelingPage';
+import labelingPageStateReducer from './labelingPage/labelingPageReducer';
 
 const rootReducer = combineReducers({
   cameras: camerasReducer,
